ciberforence/requirements.txt
requests
python-docx
fastapi
uvicorn
